Backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
pymongo==4.12.1
python-dotenv==1.1.0
bcrypt==4.3.0
pyjwt==2.10.1
dnspython==2.7.0
sentence-transformers==4.1.0
ChromaDB==1.0.8
PyPDF2==3.0.1
python-docx==1.0.1
tiktoken==0.5.1
pandas==2.2.0
passlib==1.7.4
pdfplumber==0.10.3
fastapi==0.115.9
uvicorn==0.34.2
pydantic==2.11.4
python-multipart==0.0.20
google-generativeai==0.8.5
openpyxl==3.1.5
pillow==11.2.1 